+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 - load sprites from json assets
 - game with render-loop, keyboard input, taps
 - game widget, sprite widget, animation widget
-- sprite, animation and keyboard examples
+- sprite, animation, keyboard and benchmark examples